requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto3==1.16.46
botocore==1.19.46
certifi==2020.6.20
cffi==1.14.2
colorama==0.4.3
Django==3.0.3
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-storages==1.11.1
Faker==4.1.2
gunicorn==20.0.4
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
